Fixed so that the questions are ordered correctly. Renamed the converter versions file so as not to confuse with version numbers of the applications.
